Fix Toast constraint issue
For some reason, the top margin was way the heck down low,
which pushed the text out of its parent view.

For some reason, the bottom margin constraint got disabled.

That reason seems to have something to do with the "Safe Area Relative
Margins" setting, because disabling that fixed it.